compartmentalizing files
