+++ shape.hh
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <raylib.h>
 #include <vector>
